Refactor RESCAN out of dos_programs
