+++ src/dos/meson.build
@@ -31,6 +31,7 @@ libdos_sources = files([
   'program_mem.cpp',
   'program_mount_common.cpp',
   'program_mount.cpp',
+  'program_rescan.cpp',
 ])
 
 libdos = static_library('dos', libdos_sources,
