+++ src/dos/program_rescan.cpp
@@ -1,0 +1,56 @@
+/*
+ *  SPDX-License-Identifier: GPL-2.0-or-later
+ *
+ *  Copyright (C) 2020-2021  The DOSBox Staging Team
+ *
+ *  This program is free software; you can redistribute it and/or modify
+ *  it under the terms of the GNU General Public License as published by
+ *  the Free Software Foundation; either version 2 of the License, or
+ *  (at your option) any later version.
+ *
+ *  This program is distributed in the hope that it will be useful,
+ *  but WITHOUT ANY WARRANTY; without even the implied warranty of
+ *  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ *  GNU General Public License for more details.
+ *
+ *  You should have received a copy of the GNU General Public License along
+ *  with this program; if not, write to the Free Software Foundation, Inc.,
+ *  51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
+
+#include "program_rescan.h"
+
+void RESCAN::Run(void)
+{
+	bool all = false;
+
+	Bit8u drive = DOS_GetDefaultDrive();
+
+	if (cmd->FindCommand(1,temp_line)) {
+		//-A -All /A /All
+		if (temp_line.size() >= 2 &&
+		    (temp_line[0] == '-' || temp_line[0] == '/') &&
+		    (temp_line[1] == 'a' || temp_line[1] == 'A')) {
+			all = true;
+		} else if (temp_line.size() == 2 && temp_line[1] == ':') {
+			lowcase(temp_line);
+			drive  = temp_line[0] - 'a';
+		}
+	}
+	// Get current drive
+	if (all) {
+		for (Bitu i =0; i<DOS_DRIVES; i++) {
+			if (Drives[i]) Drives[i]->EmptyCache();
+		}
+		WriteOut(MSG_Get("PROGRAM_RESCAN_SUCCESS"));
+	} else {
+		if (drive < DOS_DRIVES && Drives[drive]) {
+			Drives[drive]->EmptyCache();
+			WriteOut(MSG_Get("PROGRAM_RESCAN_SUCCESS"));
+		}
+	}
+}
+
+void RESCAN_ProgramStart(Program **make) {
+	*make=new RESCAN;
+}
